Implementing parsing of JSON response and added DateTime format helper method
Implemented datetime parsing from json response

minor updates to text format

+++ lib/model/weather_forecast_model.dart
@@ -99,7 +99,7 @@ class Current {
     uvi = json['uvi'];
     clouds = json['clouds'];
     visibility = json['visibility'];
-    windSpeed = json['wind_speed'];
+    windSpeed = json['wind_speed'].toDouble();
     windDeg = json['wind_deg'];
     if (json['weather'] != null) {
       weather = new List<Weather>();
